TPASPKT-1275 Remove DEFAULT_VERSIONS
Currently default version is required only for ODF Formula file, which
is the only file format that file-scraper detects without detecting the
format version.

However, importing ODF Formula files fails anyways, because we try to
grade the file before setting the default version. It is probably better
to remove the DEFAULT_VERSIONS dictionary, and fix the problem in
file-scraper.

See issue TPASPKT-1278 for more information.